detect/analyzer: export dump function from keyword
